SDL_Delay is notoriously innaccurate
I always prefer native API's when it comes to sensitive things like Sleeping.  Nanosleep and Sleep are far more accurate than SDL's implementation and are more performant.

+++ src/engine/p_pspr.h
@@ -65,7 +65,7 @@ typedef enum {
 
 typedef struct {
 	state_t* state;    // a NULL state means not active
-	int         tics;
+	intptr_t    tics;
 	fixed_t     sx;
 	fixed_t     sy;
 	int         alpha;  // [d64] for rendering
